feat: third and fourth use case along with validation, tests and bootstrapping the data
- Created a new abstract class, BookingService, with a single abstract
method, getAvailableSeats, which takes a list of seats and a list of
bookings as input and returns a list of seats that are available for
booking.
- Created a new implementation of BookingService,
BookingServiceImpl, which overrides the getAvailableSeats method. The
implementation uses a stream to get all the booked seats and stores them
in a set, then iterates over the list of seats and filters out any seats
that are booked or have booked seats in the surrounding seats.
- Added a new method, searchAvailableSeats, to the abstract class
ScreeningService. The method takes a screening id as input and returns
an optional ScreeningBookingInfoDTO object.
- Added a new implementation
of ScreeningService, ScreeningServiceImpl, which overrides the
searchAvailableSeats method. The implementation retrieves the requested
screening and uses the BookingService to get a list of available seats
for that screening. It then maps the resulting data to a
ScreeningBookingInfoDTO object and returns it wrapped in an optional.
- Created a new ScreeningBootstrap component to initialize some test data
for screenings. The component creates 100 screenings and saves them to
the database, using movies and screening rooms from the database. The
screenings are spaced out by 12 hours and start at the current time.
-Created a new ScreeningRoomBootstrap component to initialize some test
data for screening rooms. The component creates 10 screening rooms and
saves them to the database.
- Created a new SeatBootstrap component to
initialize some test data for seats. The component creates seats for all
the screening rooms in the database, with the number of rows and columns
specified for each screening room.
- Created a new MovieBootstrap
component to initialize some test data for movies. The component creates
5 movies and saves them to the database.
- Added a new field, duration,
to the ScreeningViewInfoDTO class.
- Added a new field, duration, to the
Screening entity class.
- Updated the ScreeningViewInfoMapper class to
map the duration field when converting between Screening and
ScreeningViewInfoDTO objects.
- Updated the
MoviesScreeningsSearchByTimeUseCase test class to use mock data and
assertions to test the searchByTime method of the ScreeningService.
- Added a new test method to the MoviesScreeningsSearchByTimeUseCase test
class to test the searchAvailableSeats method of the
ScreeningService.
Created a new class called SearchAvailableSeatsUseCase
that has a method called execute that takes in a UUID as a parameter and
returns a ScreeningBookingInfoDTO object. This method uses the
ScreeningService to search for available seats and the SeatMapper to map
the resulting list of Seat objects to a list of SeatDTO objects.
Added a test for the SearchAvailableSeatsUseCase class that mock the
dependencies and verifies the correct behavior of the execute method.
- Added a new DTO class called ScreeningBookingInfoDTO with fields for the
screening ID, movie title, duration, start time, and a list of SeatDTO
objects.
- Added a new DTO class called SeatDTO with fields for the seat
ID and seat number.
- Created a new mapper interface called SeatMapper
using the Mapstruct library and implemented it in a class called
SeatMapperImpl.
- Added a method to the SeatMapper interface called
toDto that takes in a Seat object and returns a SeatDTO object and added
an implementation for this method in the SeatMapperImpl class.
- Added a method to the SeatMapper interface called toEntity that takes in a
SeatDTO object and returns a Seat object and added an implementation for
this method in the SeatMapperImpl class.